Move MCP commands out of the tour popup into /help/mcp with copy buttons
- LaunchWelcome MCP step is now just a one-line blurb plus the "Open docs"
  button — the popup is too cramped to hold real commands. Drops the inline
  CopyableCommand widget and the hardcoded command constants.
- /help/mcp wraps each command block in a CopyBlock with a corner copy/check
  toggle, so users can grab the Claude Code, Codex, or Claude Desktop
  snippets in one click.
