+++ dali-api/app/components/LaunchWelcome.tsx
@@ -9,45 +9,11 @@ import {
   X as XIcon,
   Check,
   Bot,
-  Copy,
 } from "lucide-react";
 import { Modal } from "./Modal";
 
 const DONE_KEY = "dalios-launch-welcome-seen-v1";
 const STEP_KEY = "dalios-launch-tour-step-v1";
-
-const CLAUDE_MCP_COMMAND =
-  "claude mcp add --transport http dalios https://os.dali.dartmouth.edu/mcp";
-const CODEX_MCP_COMMAND =
-  "codex mcp add dalios https://os.dali.dartmouth.edu/mcp";
-
-function CopyableCommand({ command }: { command: string }) {
-  const [copied, setCopied] = useState(false);
-  function copy() {
-    try {
-      navigator.clipboard.writeText(command);
-      setCopied(true);
-      window.setTimeout(() => setCopied(false), 1500);
-    } catch {
-      // navigator.clipboard unavailable (insecure context, etc) — give up silently
-    }
-  }
-  return (
-    <div className="flex items-center gap-2 rounded-md bg-zinc-900 px-2.5 py-2 text-xs font-mono text-zinc-100">
-      <code className="flex-1 truncate" title={command}>
-        {command}
-      </code>
-      <button
-        type="button"
-        onClick={copy}
-        className="text-zinc-400 hover:text-white flex-shrink-0"
-        aria-label="Copy command"
-      >
-        {copied ? <Check className="w-3.5 h-3.5" /> : <Copy className="w-3.5 h-3.5" />}
-      </button>
-    </div>
-  );
-}
 
 type Phase = "modal" | "card" | "done";
 
@@ -137,24 +103,8 @@ const STEPS: TourStep[] = [
     eyebrow: "Connect any AI",
     cta: (
       <>
-        <p>
-          Connect any AI assistant to the <strong>DALI OS MCP</strong> and let
-          it read your DALI data for you.
-        </p>
-        <div className="mt-2 space-y-2">
-          <div>
-            <div className="text-[10px] font-semibold uppercase tracking-wide text-muted-foreground mb-1">
-              Claude Code
-            </div>
-            <CopyableCommand command={CLAUDE_MCP_COMMAND} />
-          </div>
-          <div>
-            <div className="text-[10px] font-semibold uppercase tracking-wide text-muted-foreground mb-1">
-              Codex
-            </div>
-            <CopyableCommand command={CODEX_MCP_COMMAND} />
-          </div>
-        </div>
+        Connect any AI assistant to the <strong>DALI OS MCP</strong> and let
+        it read your DALI data for you.
       </>
     ),
     action: {
